reorganize lib BabylonImgui layout
