+++ src/BabylonImGui/Readme.md
@@ -1,0 +1,9 @@
+# BabylonImGui
+
+This library groups :
+
+* a widget based on [ImGui](https://github.com/ocornut/imgui/) that is able to display a 3d scene (`imgui_scene_widget`)
+* the OpenGl backend for BabylonCpp (gl_rendering_context)
+* an implementation of BABYLON::ICanvas that is able to render in an OpenGl framebuffer, i.e an offscreen texture (`framebuffer_canvas`)
+* the babylon inspector widgets 
+
